requirements.txt
anthropic==0.40.0
openai==1.54.3
google-generativeai==0.8.3
gradio==5.46.1
httpx==0.27.2
tiktoken
python-dotenv
pymongo